+++ packages/prettier-plugin-apex/tests/with_multiline_map/jsfmt.spec.ts
@@ -1,0 +1,6 @@
+import { fileURLToPath } from "url";
+
+runSpec(fileURLToPath(new URL(".", import.meta.url)), ["apex"], {
+  apexMapMultiline: true,
+});
+